Course Ownership

+++ controllers/courses.js
@@ -48,6 +48,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access    Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
@@ -55,6 +56,12 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
         return next(
             new ErrorResponse(`No bootcamp with the id of ${req.params.id}`),
             404
+        );
+    }
+
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`, 401)
         );
     }
 
@@ -79,6 +86,12 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
         return next(new ErrorResponse(`Course not found with id of ${req.params.id}`, 404));
     }
 
+    if(course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to update course ${course._id}`, 401)
+        );
+    }
+
     res.status(200).json({ success: true, data: course });
 });
 
@@ -90,6 +103,12 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
 
     if(!course) {
         return next(new ErrorResponse(`Course not found with id of ${req.params.id}`, 404));
+    }
+
+    if(course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User ${req.user.id} is not authorized to delete course ${course._id}`, 401)
+        );
     }
 
     await course.deleteOne();
